ecommerce_test/celery.py
import os

import dotenv

dotenv.load_dotenv()

from celery import Celery

# Set the default Django settings module for the 'celery' program.
os.environ.setdefault("DJANGO_SETTINGS_MODULE", "ecommerce_test.settings")

app = Celery("ecommerce_test")

# Using a string here means the worker doesn't have to serialize
# the configuration object to child processes.
# - namespace='CELERY' means all celery-related configuration keys
#   should have a `CELERY_` prefix.
app.config_from_object("django.conf:settings", namespace="CELERY")

# Load task modules from all registered Django apps.
app.autodiscover_tasks()

app.conf.broker_url = os.getenv("REDIS_URL", "redis://redis:6379/0")
